+++ presets/next.js
@@ -10,7 +10,7 @@ import nextConfig from '../rules/next/next.js';
 import a11yConfig from '../rules/accessibility/a11y.js';
 import prettierConfig from '../rules/formatting/prettier.js';
 
-export default tseslint.config([
+export default [
 	{
 		ignores: [
 			'dist/**',
@@ -46,8 +46,6 @@ export default tseslint.config([
 			parserOptions: {
 				ecmaFeatures: { jsx: true },
 				sourceType: 'module',
-				project: true,
-				tsconfigRootDir: process.cwd(),
 			},
 		},
 		settings: {
@@ -74,7 +72,7 @@ export default tseslint.config([
 			...nextConfig.plugins,
 		},
 		languageOptions: {
-			ecmaVersion: 2020,
+			ecmaVersion: 'latest',
 			globals: browserNodeGlobals,
 			parserOptions: {
 				ecmaFeatures: { jsx: true },
@@ -104,4 +102,4 @@ export default tseslint.config([
 			'react-refresh/only-export-components': 'off',
 		},
 	},
-]);
+];